+++ core/src/main/java/org/carbondata/core/reader/ThriftReader.java
@@ -89,7 +89,7 @@ public class ThriftReader {
    * This method will set the position of stream from where data has to be read
    */
   public void setReadOffset(long bytesToSkip) throws IOException {
-    if(dataInputStream.skip(bytesToSkip) != bytesToSkip) {
+    if (dataInputStream.skip(bytesToSkip) != bytesToSkip) {
       throw new IOException("It doesn't set the offset properly");
     }
   }
